src/app/shared/components/spinner/spinner.component.ts
import { Component, OnInit } from '@angular/core';
import { SpinnerService } from '../../../core/services/spinner.service';

@Component({
  selector: 'app-spinner',
  templateUrl: './spinner.component.html',
  styleUrls: ['./spinner.component.scss']
})
export class SpinnerComponent implements OnInit {
  isLoading = false;

  constructor(private spinnerService: SpinnerService) {}

  ngOnInit(): void {
    this.spinnerService.isLoading.subscribe(
      isLoading => {
        this.isLoading = isLoading;
      },
      error => {
        this.isLoading = false;
      }
    );
  }
}
